+++ src/clients/rpc_tls_client.h
@@ -4,7 +4,7 @@
 
 #include "ccf/serdes.h"
 #include "http/http_builder.h"
-#include "http/http_consts.h"
+#include "ccf/http_consts.h"
 #include "http/http_parser.h"
 #include "tls_client.h"
