+++ src/endpoints/common_endpoint_registry.cpp
@@ -6,10 +6,10 @@
 #include "ccf/common_auth_policies.h"
 #include "ccf/ds/nonstd.h"
 #include "ccf/historical_queries_adapter.h"
+#include "ccf/http_consts.h"
 #include "ccf/http_query.h"
 #include "ccf/json_handler.h"
 #include "enclave/node_context.h"
-#include "http/http_consts.h"
 #include "node/rpc/call_types.h"
 #include "node/rpc/serialization.h"
 #include "node/tx_receipt.h"
