+++ src/http/http_jwt.h
@@ -5,7 +5,7 @@
 #include "ccf/crypto/base64.h"
 #include "crypto/hash.h"
 #include "crypto/verifier.h"
-#include "http_consts.h"
+#include "ccf/http_consts.h"
 #include "http_parser.h"
 
 #define FMT_HEADER_ONLY
