+++ src/http/http_sig.h
@@ -3,10 +3,10 @@
 #pragma once
 
 #include "ccf/crypto/base64.h"
+#include "ccf/http_consts.h"
 #include "ccf/service/signed_req.h"
 #include "crypto/hash.h"
 #include "crypto/key_pair.h"
-#include "http_consts.h"
 #include "http_parser.h"
 
 #define FMT_HEADER_ONLY
